Decouple source generation from NHapi.Base * Create new project and move source generating source code to it
