+++ Hl7Models.sln
@@ -19,8 +19,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHapi.NUnit", "tests\NHapi.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHapi.Model.V25", "src\NHapi.Model.V25\NHapi.Model.V25.csproj", "{BE6CDFAD-560E-44CC-B23B-87252DD4392C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ModelGenerator", "src\ModelGenerator\ModelGenerator.csproj", "{EBB8E9C9-7FF2-49C4-96EF-C5B87E4C2617}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHapi.Model.V22_ZSegments", "src\NHapi.Model.V22_ZSegments\NHapi.Model.V22_ZSegments.csproj", "{FFECB4DE-C192-4F95-9DC7-9D0A815D250A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHapi.Model.V251", "src\NHapi.Model.V251\NHapi.Model.V251.csproj", "{616C84EF-BE12-47A8-A524-D8230D572163}"
@@ -86,10 +84,6 @@ Global
 		{BE6CDFAD-560E-44CC-B23B-87252DD4392C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BE6CDFAD-560E-44CC-B23B-87252DD4392C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BE6CDFAD-560E-44CC-B23B-87252DD4392C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EBB8E9C9-7FF2-49C4-96EF-C5B87E4C2617}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EBB8E9C9-7FF2-49C4-96EF-C5B87E4C2617}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EBB8E9C9-7FF2-49C4-96EF-C5B87E4C2617}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EBB8E9C9-7FF2-49C4-96EF-C5B87E4C2617}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FFECB4DE-C192-4F95-9DC7-9D0A815D250A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FFECB4DE-C192-4F95-9DC7-9D0A815D250A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FFECB4DE-C192-4F95-9DC7-9D0A815D250A}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -135,7 +129,6 @@ Global
 		{4B3417BA-0090-4FA0-B04B-BB19563FC11C} = {9FC71692-FA33-4213-8323-E53688BCA633}
 		{7C7C99F0-672B-4DB1-A125-5ED7A3ADF95E} = {65FE0FF9-296E-4C9B-A670-05D7AC663569}
 		{BE6CDFAD-560E-44CC-B23B-87252DD4392C} = {9FC71692-FA33-4213-8323-E53688BCA633}
-		{EBB8E9C9-7FF2-49C4-96EF-C5B87E4C2617} = {9FC71692-FA33-4213-8323-E53688BCA633}
 		{FFECB4DE-C192-4F95-9DC7-9D0A815D250A} = {9FC71692-FA33-4213-8323-E53688BCA633}
 		{616C84EF-BE12-47A8-A524-D8230D572163} = {9FC71692-FA33-4213-8323-E53688BCA633}
 		{A21B73C4-D18E-4BBA-BE6C-FD52A1F5F6C9} = {9FC71692-FA33-4213-8323-E53688BCA633}
