+++ src/NHapi.Base/Parser/DefaultModelClassFactory.cs
@@ -8,7 +8,6 @@ using System.Collections.Generic;
 using System.Threading;
 using NHapi.Base;
 using NHapi.Base.Model;
-using NHapi.Base.SourceGeneration;
 using NHapi.Base.Log;
 using NHapi.Base.Model.Configuration;
