+++ src/NHapi.Base/SupportClass.cs
@@ -11,9 +11,6 @@
 
 using System;
 using System.Collections;
-using System.Data;
-using System.Data.Common;
-using System.Data.Odbc;
 using System.Globalization;
 using System.IO;
 using System.Reflection;
@@ -1971,739 +1968,6 @@ namespace NHapi.Base
 
 		/*******************************/
 
-		public class TransactionManager
-		{
-			public static ConnectionHashTable manager = new ConnectionHashTable();
-
-			public class ConnectionHashTable : Hashtable
-			{
-				public DbCommand CreateStatement(OdbcConnection connection)
-				{
-					DbCommand command = connection.CreateCommand();
-					DbTransaction transaction;
-					if (this[connection] != null)
-					{
-						ConnectionProperties Properties = ((ConnectionProperties) this[connection]);
-						transaction = Properties.Transaction;
-						command.Transaction = transaction;
-						command.CommandTimeout = 0;
-					}
-					else
-					{
-						ConnectionProperties TempProp = new ConnectionProperties();
-						TempProp.AutoCommit = true;
-						TempProp.TransactionLevel = 0;
-						command.Transaction = TempProp.Transaction;
-						command.CommandTimeout = 0;
-						Add(connection, TempProp);
-					}
-					return command;
-				}
-
-				public void Commit(OdbcConnection connection)
-				{
-					if (this[connection] != null && !((ConnectionProperties) this[connection]).AutoCommit)
-					{
-						ConnectionProperties Properties = ((ConnectionProperties) this[connection]);
-						DbTransaction transaction = Properties.Transaction;
-						transaction.Commit();
-						if (Properties.TransactionLevel == 0)
-							Properties.Transaction = connection.BeginTransaction();
-						else
-							Properties.Transaction = connection.BeginTransaction(Properties.TransactionLevel);
-					}
-				}
-
-				public void RollBack(OdbcConnection connection)
-				{
-					if (this[connection] != null && !((ConnectionProperties) this[connection]).AutoCommit)
-					{
-						ConnectionProperties Properties = ((ConnectionProperties) this[connection]);
-						DbTransaction transaction = Properties.Transaction;
-						transaction.Rollback();
-						if (Properties.TransactionLevel == 0)
-							Properties.Transaction = connection.BeginTransaction();
-						else
-							Properties.Transaction = connection.BeginTransaction(Properties.TransactionLevel);
-					}
-				}
-
-				public void SetAutoCommit(OdbcConnection connection, bool boolean)
-				{
-					if (this[connection] != null)
-					{
-						ConnectionProperties Properties = ((ConnectionProperties) this[connection]);
-						if (Properties.AutoCommit != boolean)
-						{
-							Properties.AutoCommit = boolean;
-							if (!boolean)
-							{
-								if (Properties.TransactionLevel == 0)
-									Properties.Transaction = connection.BeginTransaction();
-								else
-									Properties.Transaction = connection.BeginTransaction(Properties.TransactionLevel);
-							}
-							else
-							{
-								DbTransaction transaction = Properties.Transaction;
-								if (transaction != null)
-								{
-									transaction.Commit();
-								}
-							}
-						}
-					}
-					else
-					{
-						ConnectionProperties TempProp = new ConnectionProperties();
-						TempProp.AutoCommit = boolean;
-						TempProp.TransactionLevel = 0;
-						if (!boolean)
-							TempProp.Transaction = connection.BeginTransaction();
-						Add(connection, TempProp);
-					}
-				}
-
-				public DbCommand PrepareStatement(OdbcConnection connection, String sql)
-				{
-					DbCommand command = CreateStatement(connection);
-					command.CommandText = sql;
-					command.CommandTimeout = 0;
-					return command;
-				}
-
-				public DbCommand PrepareCall(OdbcConnection connection, String sql)
-				{
-					DbCommand command = CreateStatement(connection);
-					command.CommandText = sql;
-					command.CommandTimeout = 0;
-					return command;
-				}
-
-				public void SetTransactionIsolation(OdbcConnection connection, int level)
-				{
-					ConnectionProperties Properties;
-					if (level == (int) IsolationLevel.ReadCommitted)
-						SetAutoCommit(connection, false);
-					else if (level == (int) IsolationLevel.ReadUncommitted)
-						SetAutoCommit(connection, false);
-					else if (level == (int) IsolationLevel.RepeatableRead)
-						SetAutoCommit(connection, false);
-					else if (level == (int) IsolationLevel.Serializable)
-						SetAutoCommit(connection, false);
-
-					if (this[connection] != null)
-					{
-						Properties = ((ConnectionProperties) this[connection]);
-						Properties.TransactionLevel = (IsolationLevel) level;
-					}
-					else
-					{
-						Properties = new ConnectionProperties();
-						Properties.AutoCommit = true;
-						Properties.TransactionLevel = (IsolationLevel) level;
-						Add(connection, Properties);
-					}
-				}
-
-				public int GetTransactionIsolation(OdbcConnection connection)
-				{
-					if (this[connection] != null)
-					{
-						ConnectionProperties Properties = ((ConnectionProperties) this[connection]);
-						if (Properties.TransactionLevel != 0)
-							return (int) Properties.TransactionLevel;
-						else
-							return 2;
-					}
-					else
-						return 2;
-				}
-
-				public bool GetAutoCommit(OdbcConnection connection)
-				{
-					if (this[connection] != null)
-						return ((ConnectionProperties) this[connection]).AutoCommit;
-					else
-						return true;
-				}
-
-				/// <summary>
-				/// Sets the value of a parameter using any permitted object.  The given argument object will be converted to the
-				/// corresponding SQL type before being sent to the database.
-				/// </summary>
-				/// <param name="command">Command object to be changed.</param>
-				/// <param name="parameterIndex">One-based index of the parameter to be set.</param>
-				/// <param name="parameter">The object containing the input parameter value.</param>
-				public void SetValue(DbCommand command, int parameterIndex, Object parameter)
-				{
-					if (command.Parameters.Count < parameterIndex)
-						command.Parameters.Add(command.CreateParameter());
-					command.Parameters[parameterIndex - 1].Value = parameter;
-				}
-
-				/// <summary>
-				/// Sets a parameter to SQL NULL.
-				/// </summary>
-				/// <param name="command">Command object to be changed.</param>
-				/// <param name="parameterIndex">One-based index of the parameter to be set.</param>
-				/// <param name="targetSqlType">The SQL type to be sent to the database.</param>
-				public void SetNull(DbCommand command, int parameterIndex, int sqlType)
-				{
-					if (command.Parameters.Count < parameterIndex)
-						command.Parameters.Add(command.CreateParameter());
-					command.Parameters[parameterIndex - 1].Value = Convert.DBNull;
-					command.Parameters[parameterIndex - 1].DbType = (DbType) sqlType;
-				}
-
-				/// <summary>
-				/// Sets the value of a parameter using an object.  The given argument object will be converted to the
-				/// corresponding SQL type before being sent to the database.
-				/// </summary>
-				/// <param name="command">Command object to be changed.</param>
-				/// <param name="parameterIndex">One-based index of the parameter to be set.</param>
-				/// <param name="parameter">The object containing the input parameter value.</param>
-				/// <param name="targetSqlType">The SQL type to be sent to the database.</param>
-				public void SetObject(DbCommand command, int parameterIndex, Object parameter, int targetSqlType)
-				{
-					if (command.Parameters.Count < parameterIndex)
-						command.Parameters.Add(command.CreateParameter());
-					command.Parameters[parameterIndex - 1].Value = parameter;
-					command.Parameters[parameterIndex - 1].DbType = (DbType) targetSqlType;
-				}
-
-				/// <summary>
-				/// Sets the value of a parameter using an object.  The given argument object will be converted to the
-				/// corresponding SQL type before being sent to the database.
-				/// </summary>
-				/// <param name="command">Command object to be changed.</param>
-				/// <param name="parameterIndex">One-based index of the parameter to be set.</param>
-				/// <param name="parameter">The object containing the input parameter value.</param>
-				public void SetObject(DbCommand command, int parameterIndex, Object parameter)
-				{
-					if (command.Parameters.Count < parameterIndex)
-						command.Parameters.Add(command.CreateParameter());
-					command.Parameters[parameterIndex - 1].Value = parameter;
-				}
-
-				/// <summary>
-				/// This method is for such prepared statements verify if the Conection is autoCommit for assing the transaction to the command.
-				/// </summary>
-				/// <param name="command">The command to be tested.</param>
-				/// <returns>The number of rows afected.</returns>
-				public int ExecuteUpdate(DbCommand command)
-				{
-					if (!(((ConnectionProperties) this[command.Connection]).AutoCommit))
-					{
-						command.Transaction = ((ConnectionProperties) this[command.Connection]).Transaction;
-						return command.ExecuteNonQuery();
-					}
-					else
-						return command.ExecuteNonQuery();
-				}
-
-				/// <summary>
-				/// This method Closes the connection, and if the property of autocommit is true make the comit operation
-				/// </summary>
-				/// <param name="command"> The command to be closed</param>		
-				public void Close(OdbcConnection Connection)
-				{
-					if ((this[Connection] != null) && !(((ConnectionProperties) this[Connection]).AutoCommit))
-					{
-						Commit(Connection);
-					}
-					Connection.Close();
-				}
-
-				private class ConnectionProperties
-				{
-					public bool AutoCommit;
-					public DbTransaction Transaction;
-					public IsolationLevel TransactionLevel;
-				}
-			}
-		}
-
-		/*******************************/
-
-		/// <summary>
-		/// Manager for the connection with the database
-		/// </summary>
-		public class OdbcDBSchema
-		{
-			private DataTable schemaData = null;
-			private OdbcConnection Connection;
-			private ConnectionState ConnectionState;
-
-			/// <summary>
-			/// Constructs a new member with the provided connection
-			/// </summary>
-			/// <param name="Connection">The connection to assign to the new member</param>
-			public OdbcDBSchema(OdbcConnection Connection)
-			{
-				this.Connection = Connection;
-			}
-
-			/// <summary>
-			/// Gets the Driver name of the connection
-			/// </summary>
-			public String DriverName
-			{
-				get
-				{
-					String result = "";
-					OpenConnection();
-					result = Connection.Driver;
-					CloseConnection();
-					return result;
-				}
-			}
-
-			/// <summary>
-			/// Opens the connection
-			/// </summary>
-			private void OpenConnection()
-			{
-				ConnectionState = Connection.State;
-				Connection.Close();
-				Connection.Open();
-				schemaData = null;
-			}
-
-			/// <summary>
-			/// Closes the connection
-			/// </summary>
-			private void CloseConnection()
-			{
-				if (ConnectionState == ConnectionState.Open)
-					Connection.Close();
-			}
-
-			/// <summary>
-			/// Gets the info of the row
-			/// </summary>
-			/// <param name="filter">Filter to apply to the row</param>
-			/// <param name="RowName">The row from which to obtain the filter</param>
-			/// <returns>A new String with the info from the row</returns>
-			private String GetMaxInfo(String filter, String RowName)
-			{
-				String result = "";
-				schemaData = null;
-				OpenConnection();
-				schemaData = Connection.GetSchema("DbInfoLiterals", null);
-				foreach (DataRow DataRow in schemaData.Rows)
-				{
-					if (DataRow["LiteralName"].ToString() == filter)
-					{
-						result = DataRow[RowName].ToString();
-						break;
-					}
-				}
-				CloseConnection();
-				return result;
-			}
-
-			/// <summary>
-			/// Gets the catalogs from the database to which it is connected
-			/// </summary>
-			public DataTable Catalogs
-			{
-				get
-				{
-					OpenConnection();
-					schemaData = Connection.GetSchema("Catalogs", null);
-					CloseConnection();
-					return schemaData;
-				}
-			}
-
-			/// <summary>
-			/// Gets the OleDBConnection for the current member
-			/// </summary>
-			/// <returns></returns>
-			public OdbcConnection GetConnection()
-			{
-				return Connection;
-			}
-
-			/// <summary>
-			/// Gets a description of the stored procedures available
-			/// </summary>
-			/// <param name="catalog">The catalog from which to obtain the procedures</param>
-			/// <param name="schemaPattern">Schema pattern, retrieves those without the schema</param>
-			/// <param name="procedureNamePattern">a procedure name pattern</param>
-			/// <returns>each row but withing a procedure description</returns>
-			public DataTable GetProcedures(String catalog, String schemaPattern, String procedureNamePattern)
-			{
-				OpenConnection();
-				schemaData = Connection.GetSchema("Procedures",
-					new [] {catalog, schemaPattern, procedureNamePattern, null});
-				CloseConnection();
-				return schemaData;
-			}
-
-			/// <summary>
-			/// Gets a collection of the descriptions of the stored procedures parameters and result columns
-			/// </summary>
-			/// <param name="catalog">Retrieves those without a catalog</param>
-			/// <param name="schemaPattern">Schema pattern, retrieves those without the schema</param>
-			/// <param name="procedureNamePattern">a procedure name pattern</param>
-			/// <param name="columnNamePattern">a columng name patterm</param>
-			/// <returns>Each row but withing a procedure description or column</returns>
-			public DataTable GetProcedureColumns(String catalog, String schemaPattern, String procedureNamePattern,
-				String columnNamePattern)
-			{
-				OpenConnection();
-				schemaData = Connection.GetSchema("Procedure_Parameters",
-					new [] {catalog, schemaPattern, procedureNamePattern, columnNamePattern});
-				CloseConnection();
-				return schemaData;
-			}
-
-			/// <summary>
-			/// Gets a description of the tables available for the catalog
-			/// </summary>
-			/// <param name="catalog">A catalog, retrieves those without a catalog</param>
-			/// <param name="schemaPattern">Schema pattern, retrieves those without the schema</param>
-			/// <param name="tableNamePattern">A table name pattern</param>
-			/// <param name="types">a list of table types to include</param>
-			/// <returns>Each row</returns>
-			public DataTable GetTables(String catalog, String schemaPattern, String tableNamePattern, String[] types)
-			{
-				OpenConnection();
-				schemaData = Connection.GetSchema("Tables",
-					new [] {catalog, schemaPattern, tableNamePattern, types[0]});
-				if (types != null)
-				{
-					for (int i = 1; i < types.Length; i++)
-					{
-						DataTable temp_Table = Connection.GetSchema("Tables",
-							new [] {catalog, schemaPattern, tableNamePattern, types[i]});
-						for (int j = 0; j < temp_Table.Rows.Count; j++)
-						{
-							schemaData.ImportRow(temp_Table.Rows[j]);
-						}
-					}
-				}
-				CloseConnection();
-				return schemaData;
-			}
-
-			/// <summary>
-			/// Gets a description of the table rights
-			/// </summary>
-			/// <param name="catalog">A catalog, retrieves those without a catalog</param>
-			/// <param name="schemaPattern">Schema pattern, retrieves those without the schema</param>
-			/// <param name="tableNamePattern">A table name pattern</param>
-			/// <returns>A description of the table rights</returns>
-			public DataTable GetTablePrivileges(String catalog, String schemaPattern, String tableNamePattern)
-			{
-				OpenConnection();
-				schemaData = Connection.GetSchema("Table_Privileges",
-					new [] {catalog, schemaPattern, tableNamePattern});
-				CloseConnection();
-				return schemaData;
-			}
-
-			/// <summary>
-			/// Gets the table types available
-			/// </summary>
-			public DataTable TableTypes
-			{
-				get
-				{
-					OpenConnection();
-					schemaData = Connection.GetSchema("Tables", null);
-					ArrayList tableTypes = new ArrayList(schemaData.Rows.Count);
-
-					String tableType = "";
-					foreach (DataRow DataRow in schemaData.Rows)
-					{
-						tableType = DataRow[schemaData.Columns["TABLE_TYPE"]].ToString();
-						if (!(tableTypes.Contains(tableType)))
-						{
-							tableTypes.Add(tableType);
-						}
-					}
-					schemaData = new DataTable();
-					schemaData.Columns.Add("TABLE_TYPE");
-					for (int index = 0; index < tableTypes.Count; index++)
-					{
-						schemaData.Rows.Add(new Object[] {tableTypes[index]});
-					}
-					CloseConnection();
-					return schemaData;
-				}
-			}
-
-			/// <summary>
-			/// Gets a description of the table columns available
-			/// </summary>
-			/// <param name="catalog">A catalog, retrieves those without a catalog</param>
-			/// <param name="schemaPattern">Schema pattern, retrieves those without the schema</param>
-			/// <param name="tableNamePattern">A table name pattern</param>
-			/// <param name="columnNamePattern">a columng name patterm</param>
-			/// <returns>A description of the table columns available</returns>
-			public DataTable GetColumns(String catalog, String schemaPattern, String tableNamePattern, String columnNamePattern)
-			{
-				OpenConnection();
-				schemaData = Connection.GetSchema("Columns",
-					new [] {catalog, schemaPattern, tableNamePattern, columnNamePattern});
-				CloseConnection();
-				return schemaData;
-			}
-
-			/// <summary>
-			/// Gets a description of the primary keys available
-			/// </summary>
-			/// <param name="catalog">A catalog, retrieves those without a catalog</param>
-			/// <param name="schema">Schema name, retrieves those without the schema</param>
-			/// <param name="table">A table name</param>
-			/// <returns>A description of the primary keys available</returns>
-			public DataTable GetPrimaryKeys(String catalog, String schema, String table)
-			{
-				OpenConnection();
-				schemaData = Connection.GetSchema("Primary_Keys", new [] {catalog, schema, table});
-				CloseConnection();
-				return schemaData;
-			}
-
-			/// <summary>
-			/// Gets a description of the foreign keys available
-			/// </summary>
-			/// <param name="catalog">A catalog, retrieves those without a catalog</param>
-			/// <param name="schema">Schema name, retrieves those without the schema</param>
-			/// <param name="table">A table name</param>
-			/// <returns>A description of the foreign keys available</returns>
-			public DataTable GetForeignKeys(String catalog, String schema, String table)
-			{
-				OpenConnection();
-				schemaData = Connection.GetSchema("Foreign_Keys", new [] {catalog, schema, table});
-				CloseConnection();
-				return schemaData;
-			}
-
-			/// <summary>
-			/// Gets a description of the access rights for	a table columns
-			/// </summary>
-			/// <param name="catalog">A catalog, retrieves those without a catalog</param>
-			/// <param name="schema">Schema name, retrieves those without the schema</param>
-			/// <param name="table">A table name</param>
-			/// <param name="columnNamePattern">A column name patter</param>
-			/// <returns>A description of the access rights for	a table columns</returns>
-			public DataTable GetColumnPrivileges(String catalog, String schema, String table, String columnNamePattern)
-			{
-				OpenConnection();
-				schemaData = Connection.GetSchema("Column_Privileges",
-					new [] {catalog, schema, table, columnNamePattern});
-				CloseConnection();
-				return schemaData;
-			}
-
-			/// <summary>
-			/// Gets the provider version
-			/// </summary>
-			public String ProviderVersion
-			{
-				get
-				{
-					String result = "";
-					OpenConnection();
-					result = Connection.ServerVersion;
-					CloseConnection();
-					return result;
-				}
-			}
-
-			/// <summary>
-			/// Gets the default transaction isolation integer value
-			/// </summary>
-			public int DefaultTransactionIsolation
-			{
-				get
-				{
-					int result = -1;
-					OpenConnection();
-					DbTransaction Transaction = Connection.BeginTransaction();
-					result = (int) Transaction.IsolationLevel;
-					CloseConnection();
-					return result;
-				}
-			}
-
-			/// <summary>
-			/// Gets the schemata for the member
-			/// </summary>
-			public DataTable Schemata
-			{
-				get
-				{
-					OpenConnection();
-					schemaData = Connection.GetSchema("Schemata", null);
-					CloseConnection();
-					return schemaData;
-				}
-			}
-
-			/// <summary>
-			/// Gets the provider types for the member
-			/// </summary>
-			public DataTable ProviderTypes
-			{
-				get
-				{
-					OpenConnection();
-					schemaData = Connection.GetSchema("Provider_Types", null);
-					CloseConnection();
-					return schemaData;
-				}
-			}
-
-			/// <summary>
-			/// Gets the catalog separator
-			/// </summary>
-			public String CatalogSeparator
-			{
-				get { return GetMaxInfo("Catalog_Separator", "LiteralValue"); }
-			}
-
-			/// <summary>
-			/// Gets the maximum binary length permited
-			/// </summary>
-			public int MaxBinaryLiteralLength
-			{
-				get
-				{
-					String len = GetMaxInfo("Binary_Literal", "Maxlen");
-					if (len.Equals(""))
-						return 0;
-					else
-						return (Convert.ToInt32(len)*4);
-				}
-			}
-
-			/// <summary>
-			/// Gets the maximum catalog name length permited
-			/// </summary>
-			public int MaxCatalogNameLength
-			{
-				get
-				{
-					String len = GetMaxInfo("Catalog_Name", "Maxlen");
-					if (len.Equals(""))
-						return 0;
-					else
-						return (Convert.ToInt32(len));
-				}
-			}
-
-			/// <summary>
-			/// Gets the maximum character literal length permited
-			/// </summary>
-			public int MaxCharLiteralLength
-			{
-				get
-				{
-					String len = GetMaxInfo("Char_Literal", "Maxlen");
-					if (len.Equals(""))
-						return 0;
-					else
-						return (Convert.ToInt32(len)*4);
-				}
-			}
-
-			/// <summary>
-			/// Gets the maximum column name length
-			/// </summary>
-			public int MaxColumnNameLength
-			{
-				get
-				{
-					String len = GetMaxInfo("Column_Name", "Maxlen");
-					if (len.Equals(""))
-						return 0;
-					else
-						return (Convert.ToInt32(len));
-				}
-			}
-
-			/// <summary>
-			/// Gets the maximum cursor name length
-			/// </summary>
-			public int MaxCursorNameLength
-			{
-				get
-				{
-					String len = GetMaxInfo("Cursor_Name", "Maxlen");
-					if (len.Equals(""))
-						return 0;
-					else
-						return (Convert.ToInt32(len));
-				}
-			}
-
-			/// <summary>
-			/// Gets the maximum procedure name length
-			/// </summary>
-			public int MaxProcedureNameLength
-			{
-				get
-				{
-					String len = GetMaxInfo("Procedure_Name", "Maxlen");
-					if (len.Equals(""))
-						return 0;
-					else
-						return (Convert.ToInt32(len));
-				}
-			}
-
-			/// <summary>
-			/// Gets the maximum schema name length
-			/// </summary>
-			public int MaxSchemaNameLength
-			{
-				get
-				{
-					String len = GetMaxInfo("Schema_Name", "Maxlen");
-					if (len.Equals(""))
-						return 0;
-					else
-						return (Convert.ToInt32(len));
-				}
-			}
-
-			/// <summary>
-			/// Gets the maximum table name length
-			/// </summary>
-			public int MaxTableNameLength
-			{
-				get
-				{
-					String len = GetMaxInfo("Table_Name", "Maxlen");
-					if (len.Equals(""))
-						return 0;
-					else
-						return (Convert.ToInt32(len));
-				}
-			}
-
-			/// <summary>
-			/// Gets the maximum user name length
-			/// </summary>
-			public int MaxUserNameLength
-			{
-				get
-				{
-					String len = GetMaxInfo("User_Name", "Maxlen");
-					if (len.Equals(""))
-						return 0;
-					else
-						return Convert.ToInt32(len);
-				}
-			}
-		}
-
-		/*******************************/
 
 		/// <summary>
 		/// This class provides functionality not found in .NET collection-related interfaces.
