+++ tests/NHapi.NUnit/NHapi.NUnit.csproj
@@ -17,7 +17,6 @@
 
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\ModelGenerator\ModelGenerator.csproj" />
     <ProjectReference Include="..\..\src\nHapi.Base\NHapi.Base.csproj" />
     <ProjectReference Include="..\..\src\NHapi.Model.V21\NHapi.Model.V21.csproj" />
     <ProjectReference Include="..\..\src\NHapi.Model.V22\NHapi.Model.V22.csproj" />
